requirements.txt
keras
numpy
opencv-python
Pillow
